@@ -1,11 +1,9 @@
 /** @type {import('tailwindcss').Config} */
-module.exports = {
-  content: [
-    "./index.html",
-    "./src/**/*.{vue,js,ts,jsx,tsx}",
-  ],
-  theme: {
-    extend: {},
-  },
-  plugins: [],
-}
+export const content = [
+  "./index.html",
+  "./src/**/*.{vue,js,ts,jsx,tsx}",
+];
+export const theme = {
+  extend: {},
+};
+export const plugins = [];
