@@ -1,15 +1,47 @@
 import { defineConfig } from 'vite'
 import vue from '@vitejs/plugin-vue'
+import { ViteImageOptimizer } from 'vite-plugin-image-optimizer';
 
-// https://vitejs.dev/config/
 export default defineConfig({
-  plugins: [vue()],
+  plugins: [vue(), ViteImageOptimizer({
+    jpeg: {
+      quality: 80,
+    },
+    jpg: {
+      quality: 80,
+    },
+    png: {
+      quality: 80,
+    },
+    webp: {
+      quality: 80,
+      lossless: false,
+    },
+    avif: {
+      quality: 80,
+      lossless: false,
+    },
+    svg: {
+      multipass: true,
+      plugins: [
+        {
+          name: 'preset-default',
+          params: {
+            overrides: {
+              cleanupNumericValues: false,
+              removeViewBox: false,
+            },
+          },
+        },
+      ],
+    }
+  }),],
   build: {
     chunkSizeWarningLimit:1500,
     rollupOptions: {
         output:{
             manualChunks(id) {
-              if (id.includes('node_modules')) {
+              if ((id as any).includes('node_modules')) {
                 
                   return id.toString().split('node_modules/')[1].split('/')[0].toString();
               }
